+++ projects/small_fem/assembler/SystemAbstract.cpp
@@ -14,30 +14,33 @@ SystemAbstract::~SystemAbstract(void){
 
 void SystemAbstract::fixCoef(const GroupOfElement& goe, double value){
   const vector<const MElement*>& element = goe.getAll();
-  const unsigned int            nElement = goe.getNumber();
+  const size_t                  nElement = goe.getNumber();
 
-  for(unsigned int i = 0; i < nElement; i++){
-    vector<Dof>         dof = fs->getKeys(*element[i]);
-    const unsigned int nDof = dof.size();
+  for(size_t i = 0; i < nElement; i++){
+    vector<Dof>   dof = fs->getKeys(*element[i]);
+    const size_t nDof = dof.size();
 
-    for(unsigned int j = 0; j < nDof; j++)
+    for(size_t j = 0; j < nDof; j++)
       dofM->fixValue(dof[j], value);
   }
 }
 
-void SystemAbstract::dirichlet(GroupOfElement& goe,
-                               double (*f)(fullVector<double>& xyz)){
+void SystemAbstract::
+dirichlet(GroupOfElement& goe,
+          double (*f)(fullVector<double>& xyz)){
 
   // Check if Scalar Problem //
   if(!fs->isScalar())
-    throw Exception("Cannot impose Vectorial Dirichlet Conditions on a Scalar Problem");
+    throw
+      Exception("Cannot impose Vectorial Dirichlet Conditions on a Scalar Problem");
 
   // New FunctionSpace, on the Dirichlet Domain: dirFS //
   // WARNING: The support of the dirFS *MUST* have the fs Mesh
   //  --> So we have the same Dof Numbering
 
   if(&(goe.getMesh()) != &(fs->getSupport().getMesh()))
-    throw Exception("Dirichlet Domain must come from the FunctionSpace Domain's Mesh");
+    throw
+      Exception("Dirichlet Domain must come from the FunctionSpace Domain's Mesh");
 
   BasisLocal* dirBasis = BasisGenerator::generate(goe.get(0).getType(),
                                                   fs->getBasis(0).getType(),
@@ -54,23 +57,25 @@ void SystemAbstract::dirichlet(GroupOfElement& goe,
 
   // Fix This System Dofs with sysProj Solution //
   const vector<const Dof*> dof = dirFS.getAllDofs();
-  const unsigned int      nDof = dof.size();
+  const size_t            nDof = dof.size();
 
   const DofManager&        dirDofM = sysProj.getDofManager();
   const fullVector<double>& dirSol = sysProj.getSol();
 
-  for(unsigned int i = 0; i < nDof; i++)
+  for(size_t i = 0; i < nDof; i++)
     dofM->fixValue(*dof[i], dirSol(dirDofM.getGlobalId(*dof[i])));
 
   delete dirBasis;
 }
 
-void SystemAbstract::dirichlet(GroupOfElement& goe,
-                               fullVector<double> (*f)(fullVector<double>& xyz)){
+void SystemAbstract::
+dirichlet(GroupOfElement& goe,
+          fullVector<double> (*f)(fullVector<double>& xyz)){
 
   // Check if Scalar Problem //
   if(fs->isScalar())
-    throw Exception("Cannot impose Scalar Dirichlet Conditions on a Vectorial Problem");
+    throw
+      Exception("Cannot impose Scalar Dirichlet Conditions on a Vectorial Problem");
 
   // New FunctionSpace, on the Dirichlet Domain: dirFS //
   // WARNING: The support of the dirFS *MUST* have the fs Mesh
@@ -94,34 +99,34 @@ void SystemAbstract::dirichlet(GroupOfElement& goe,
 
   // Fix This System Dofs with sysProj Solution //
   const vector<const Dof*> dof = dirFS.getAllDofs();
-  const unsigned int      nDof = dof.size();
+  const size_t            nDof = dof.size();
 
   const DofManager&        dirDofM = sysProj.getDofManager();
   const fullVector<double>& dirSol = sysProj.getSol();
 
-  for(unsigned int i = 0; i < nDof; i++)
+  for(size_t i = 0; i < nDof; i++)
     dofM->fixValue(*dof[i], dirSol(dirDofM.getGlobalId(*dof[i])));
 
   delete dirBasis;
 }
 
 void SystemAbstract::assemble(linearSystemPETSc<double>& sys,
-                              unsigned int elementId,
+                              size_t elementId,
                               const GroupOfDof& group,
                               formulationPtr& term){
 
   const vector<const Dof*>& dof = group.getDof();
-  const unsigned int N = group.size();
+  const size_t N = group.size();
 
-  unsigned int dofI;
-  unsigned int dofJ;
+  size_t dofI;
+  size_t dofJ;
 
-  for(unsigned int i = 0; i < N; i++){
+  for(size_t i = 0; i < N; i++){
     dofI = dofM->getGlobalId(*(dof[i]));
 
     // If not a fixed Dof line: assemble
     if(dofI != DofManager::isFixedId()){
-      for(unsigned int j = 0; j < N; j++){
+      for(size_t j = 0; j < N; j++){
         dofJ = dofM->getGlobalId(*(dof[j]));
 
         // If not a fixed Dof
@@ -144,26 +149,26 @@ void SystemAbstract::assemble(linearSystemPETSc<double>& sys,
 
 void SystemAbstract::assemble(Mat& A,
                               Vec& b,
-                              unsigned int elementId,
+                              size_t elementId,
                               const GroupOfDof& group,
                               formulationPtr& term){
 
   const vector<const Dof*>& dof = group.getDof();
-  const unsigned int N = group.size();
+  const size_t N = group.size();
 
-  unsigned int dofI;
-  unsigned int dofJ;
+  size_t dofI;
+  size_t dofJ;
 
   //PetscInt petscI;
   //PetscInt petscJ;
   //PetscScalar petscV;
 
-  for(unsigned int i = 0; i < N; i++){
+  for(size_t i = 0; i < N; i++){
     dofI = dofM->getGlobalId(*(dof[i]));
 
     // If not a fixed Dof line: assemble
     if(dofI != DofManager::isFixedId()){
-      for(unsigned int j = 0; j < N; j++){
+      for(size_t j = 0; j < N; j++){
         dofJ = dofM->getGlobalId(*(dof[j]));
 
         // If not a fixed Dof
@@ -199,17 +204,17 @@ void SystemAbstract::sparsity(linearSystemPETSc<double>& sys,
                               const GroupOfDof& group){
 
   const vector<const Dof*>& dof = group.getDof();
-  const unsigned int N = group.size();
+  const size_t N = group.size();
 
-  unsigned int dofI;
-  unsigned int dofJ;
+  size_t dofI;
+  size_t dofJ;
 
-  for(unsigned int i = 0; i < N; i++){
+  for(size_t i = 0; i < N; i++){
     dofI = dofM->getGlobalId(*(dof[i]));
 
     // Add non fixed Dof
     if(dofI != DofManager::isFixedId()){
-      for(unsigned int j = 0; j < N; j++){
+      for(size_t j = 0; j < N; j++){
         dofJ = dofM->getGlobalId(*(dof[j]));
 
         // Add non fixed Dof
@@ -225,19 +230,19 @@ void SystemAbstract::sparsity(PetscInt* nonZero,
                               const GroupOfDof& group){
 
   const vector<const Dof*>& dof = group.getDof();
-  const unsigned int N = group.size();
+  const size_t N = group.size();
 
-  unsigned int dofI;
-  unsigned int dofJ;
+  size_t dofI;
+  size_t dofJ;
 
   // Add each column only one
 
-  for(unsigned int i = 0; i < N; i++){
+  for(size_t i = 0; i < N; i++){
     dofI = dofM->getGlobalId(*(dof[i]));
 
     // Add non fixed Dof
     if(dofI != DofManager::isFixedId()){
-      for(unsigned int j = 0; j < N; j++){
+      for(size_t j = 0; j < N; j++){
         dofJ = dofM->getGlobalId(*(dof[j]));
 
         // Add non fixed Dof
