+++ projects/small_fem/formulation/FormulationEigenFrequency.h
@@ -31,20 +31,20 @@ class FormulationEigenFrequency: public Formulation{
 
  public:
   FormulationEigenFrequency(GroupOfElement& goe,
-			    unsigned int order);
+                            size_t order);
 
   virtual ~FormulationEigenFrequency(void);
 
   virtual bool isGeneral(void) const;
 
-  virtual double weak(unsigned int dofI, unsigned int dofJ,
-                      unsigned int elementId) const;
+  virtual double weak(size_t dofI, size_t dofJ,
+                      size_t elementId) const;
 
-  virtual double weakB(unsigned int dofI, unsigned int dofJ,
-		       unsigned int elementId) const;
+  virtual double weakB(size_t dofI, size_t dofJ,
+                       size_t elementId) const;
 
-  virtual double rhs(unsigned int equationI,
-		     unsigned int elementId) const;
+  virtual double rhs(size_t equationI,
+                     size_t elementId) const;
 
   virtual const FunctionSpace& fs(void) const;
 };
