+++ projects/small_fem/formulation/FormulationPoisson.h
@@ -19,7 +19,7 @@
 class FormulationPoisson: public Formulation{
  private:
   // Source Term //
-  static const unsigned int sourceOrder;
+  static const size_t sourceOrder;
 
   // Function Space & Basis //
   FunctionSpaceScalar* fspace;
@@ -31,20 +31,20 @@ class FormulationPoisson: public Formulation{
 
  public:
   FormulationPoisson(GroupOfElement& goe,
-		     unsigned int order);
+                     size_t order);
 
   virtual ~FormulationPoisson(void);
 
   virtual bool isGeneral(void) const;
 
-  virtual double weak(unsigned int dofI, unsigned int dofJ,
-		      unsigned int elementId) const;
+  virtual double weak(size_t dofI, size_t dofJ,
+                      size_t elementId) const;
 
-  virtual double weakB(unsigned int dofI, unsigned int dofJ,
-                       unsigned int elementId) const;
+  virtual double weakB(size_t dofI, size_t dofJ,
+                       size_t elementId) const;
 
-  virtual double rhs(unsigned int equationI,
-		     unsigned int elementId) const;
+  virtual double rhs(size_t equationI,
+                     size_t elementId) const;
 
   virtual const FunctionSpace& fs(void) const;
